fix admin

+++ server/src/main/java/com/springboot/bookstore/repository/OrderDetailsRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
 
-public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Long> {
+public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Integer> {
     @Query("SELECT p.id, p.image, p.title, p.oldPrice, p.currentPrice, SUM(od.quantity) as totalQuantity " +
             "FROM OrderDetails od " +
             "JOIN od.product p " +
